Hide HashValidator implementations.
I accidentally leaked OpenSSL headers to the public interface by
including `<openssl/md5.h>` in a file indirectly included from
`client.h`. Refactor the HashValidator classes so the implementations
are hidden in their own file.
