+++ google/cloud/storage/internal/hash_validator_test.cc
@@ -15,6 +15,7 @@
 #include "google/cloud/storage/internal/hash_validator.h"
 #include "google/cloud/internal/make_unique.h"
 #include "google/cloud/status.h"
+#include "google/cloud/storage/internal/hash_validator_impl.h"
 #include "google/cloud/storage/internal/object_requests.h"
 #include "google/cloud/storage/object_metadata.h"
 #include <gmock/gmock.h>
