+++ google/cloud/storage/storage_client.bzl
@@ -49,6 +49,7 @@ storage_client_hdrs = [
     "internal/generic_object_request.h",
     "internal/generic_request.h",
     "internal/hash_validator.h",
+    "internal/hash_validator_impl.h",
     "internal/hmac_key_requests.h",
     "internal/http_response.h",
     "internal/logging_client.h",
@@ -133,6 +134,7 @@ storage_client_srcs = [
     "internal/empty_response.cc",
     "internal/format_time_point.cc",
     "internal/hash_validator.cc",
+    "internal/hash_validator_impl.cc",
     "internal/hmac_key_requests.cc",
     "internal/http_response.cc",
     "internal/logging_client.cc",
